third_party/WebKit/LayoutTests/fast/dom/DOMImplementation/createHTMLDocument-optional-title.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script>
description('This tests optional title for createHTMLDocument');

shouldBeNull('document.implementation.createHTMLDocument().querySelector("title")');
shouldBeNull('document.implementation.createHTMLDocument(undefined).querySelector("title")');

shouldBeEqualToString('document.implementation.createHTMLDocument(null).querySelector("title").textContent', 'null');
shouldBeEqualToString('document.implementation.createHTMLDocument("").querySelector("title").textContent', '');
shouldBeEqualToString('document.implementation.createHTMLDocument("x").querySelector("title").textContent', 'x');

</script>
